use vec instead of HashMap for heap

+++ src/gc_values.rs
@@ -1,4 +1,4 @@
-pub type Id = u64;
+pub type Id = usize;
 
 #[derive(Copy, Clone)]
 pub struct GcString(pub Id);
